+++ src/models/Address.php
@@ -238,10 +238,12 @@ class Address extends Model
         $labels = parent::attributeLabels();
         $labels['firstName'] = Craft::t('commerce', 'First Name');
         $labels['lastName'] = Craft::t('commerce', 'Last Name');
+        $labels['fullName'] = Craft::t('commerce', 'Full Name');
         $labels['attention'] = Craft::t('commerce', 'Attention');
         $labels['title'] = Craft::t('commerce', 'Title');
         $labels['address1'] = Craft::t('commerce', 'Address 1');
         $labels['address2'] = Craft::t('commerce', 'Address 2');
+        $labels['address3'] = Craft::t('commerce', 'Address 3');
         $labels['city'] = Craft::t('commerce', 'City');
         $labels['zipCode'] = Craft::t('commerce', 'Zip Code');
         $labels['phone'] = Craft::t('commerce', 'Phone');
@@ -253,6 +255,12 @@ class Address extends Model
         $labels['stateId'] = Craft::t('commerce', 'State');
         $labels['stateName'] = Craft::t('commerce', 'State');
         $labels['stateValue'] = Craft::t('commerce', 'State');
+        $labels['custom1'] = Craft::t('commerce', 'Custom 1');
+        $labels['custom2'] = Craft::t('commerce', 'Custom 2');
+        $labels['custom3'] = Craft::t('commerce', 'Custom 3');
+        $labels['custom4'] = Craft::t('commerce', 'Custom 4');
+        $labels['notes'] = Craft::t('commerce', 'Notes');
+        $labels['label'] = Craft::t('commerce', 'Label');
         return $labels;
     }
 
@@ -264,8 +272,36 @@ class Address extends Model
         $rules = parent::rules();
         $rules[] = [['firstName'], 'required'];
         $rules[] = [['lastName'], 'required'];
-        $rules[] = ['stateId', 'validateState', 'skipOnEmpty' => false];
-        $rules[] = ['businessTaxId', 'validateBusinessTaxId', 'skipOnEmpty' => true];
+        $rules[] = [['stateId'], 'validateState', 'skipOnEmpty' => false];
+        $rules[] = [['businessTaxId'], 'validateBusinessTaxId', 'skipOnEmpty' => true];
+
+        $rules[] = [[
+            'firstName',
+            'lastName',
+            'fullName',
+            'attention',
+            'title',
+            'address1',
+            'address2',
+            'address3',
+            'city',
+            'zipCode',
+            'phone',
+            'alternativePhone',
+            'businessName',
+            'businessId',
+            'businessTaxId',
+            'countryId',
+            'stateId',
+            'stateName',
+            'stateValue',
+            'custom1',
+            'custom2',
+            'custom3',
+            'custom4',
+            'notes',
+            'label',
+        ], 'trim'];
 
         $event = new RegisterAddressRulesEvent([
             'rules' => $rules